:boom: [Breaking] Deprecate `get_interface` method
