+++ README.md
@@ -2,7 +2,7 @@
 
 # GraiaX Playwright
 
-*适用于 Graia Project 的 Playwright 管理器*
+_适用于 Graia Project 的 Playwright 管理器_
 
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Imports: isort](https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat&labelColor=ef8336)](https://pycqa.github.io/isort/)
@@ -53,20 +53,15 @@ launart.launch_blocking()
 from creart import create
 from launart import Launart
 from graia.ariadne.util.saya import listen
-from graiax.playwright import PlaywrightBrowser
-
-# 此处代码为没有使用 Persistent Context 的示例
-# 若使用 Persistent Context 请使用 `context = launart.get_interface(PlaywrightContext)`
-# 该方法获得的对象与 playwright.async_api.BrowserContext 兼容
+from graiax.playwright import PlaywrightService
 
 
 @listen(...)
 async def function(app: Ariadne):
     launart = create(Launart)
-    browser = launart.get_interface(PlaywrightBrowser)
-    # 此处的 browser 之用法与 playwright.async_api.Browser 无异，但要注意的是下方代码的返回值为 False。
-    # `isinstance(browser, playwright.async_api.Browser)`
-    async with browser.page(  # 此 API 启用了自动上下文管理
+    pw_service = launart.get_component(PlaywrightService)
+
+    async with pw_service.page(  # 此 API 启用了自动上下文管理
         viewport={"width": 800, "height": 10},
         device_scale_factor=1.5,
     ) as page:
@@ -83,14 +78,16 @@ async def function(app: Ariadne):
 > [!NOTE]  
 > 该种用法不支持持久性上下文（Persistent Context）
 >
+> 该种用法中的新上下文，仍受到 Playwright 启动参数的影响
+>
 > 更多信息详见：<https://playwright.dev/python/docs/browser-contexts>
 
 ```python
 @listen(...)
 async def function(app: Ariadne):
     launart = create(Launart)
-    browser = launart.get_interface(PlaywrightBrowser)
-    async with browser.page(new_context=True) as page:  # 此 API 启用了自动上下文管理
+    pw_service = launart.get_component(PlaywrightService)
+    async with pw_service.page(new_context=True) as page:  # 此 API 启用了自动上下文管理
         await page.set_content("Hello World!")
         img = await page.screenshot(type="jpeg", quality=80, full_page=True, scale="device")
     ...
@@ -114,17 +111,66 @@ launart.add_service(PlaywrightService("chromium"))
 **Saya 模块中：**
 
 ```python
-from graiax.playwright import PlaywrightContext
+from graiax.playwright import PlaywrightService
 
 
 @listen(...)
 async def function(app: Ariadne):
     launart = create(Launart)
-    context = launart.get_interface(PlaywrightContext)
-    async with context.page() as page:  # 此 API 启用了自动上下文管理
-        await page.set_content("Hello World!")
-        img = await page.screenshot(type="jpeg", quality=80, full_page=True, scale="device")
+    pw_service = manager.get_component(PlaywrightService)
+    async with pw_service.context(...) as context:  # 此 API 启用了自动上下文管理
+        page = context.new_page()
+        try:
+            await page.set_content("Hello World!")
+            img = await page.screenshot(type="jpeg", quality=80, full_page=True, scale='device')
+        finally:
+            page.stop()
     ...
+```
+
+### 高级用法之三
+
+通过依赖注入来获取 `PlaywrightService`，前面的代码中获取 `PlaywrightService` 都需要通过 Launart
+的 `get_component()` 方法，你也可以通过自定义一个 BCC 的 `Dispatcher` 来实现依赖注入。
+
+编写一个 `Dispatcher`：
+
+```python
+from graia.broadcast.entities.dispatcher import BaseDispatcher
+from graia.broadcast.interfaces.dispatcher import DispatcherInterface
+
+class CustomDispatcher(BaseDispatcher):
+    @classmethod
+    async def catch(cls, interface: DispatcherInterface):
+        with contextlib.suppress(TypeError):
+            if generic_isinstance(interface.event, Service):
+                manager = Launart.current()
+                return manager.get_component(interface.annotation)
+```
+
+在启动 Launart 之前获取一个 BCC 实例并对其应用这个 `Dispatcher`：
+
+```python
+from creart import it
+from graia.broadcast import Broadcast
+
+bcc = it(Broadcast)
+bcc.finale_dispatchers.append(RedbotDispatcher)
+
+...
+
+launart.launch_blocking() # 或者是 avilla.launch()
+```
+
+之后用起来就很简单了，你可以对比一下有什么不同：
+
+```python
+from graiax.playwright import PlaywrightService
+
+@listen(...)
+async def function(app: Ariadne, pw_service: PlaywrightService):
+    async with pw_service.page(...) as page:
+        ...
 ```
 
 ## 许可证
