+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "graiax-playwright"
-version = "0.3.1"
+version = "0.4.0"
 description = "适用于 Graia 的 Playwright 管理器"
 authors = [
     { name = "BlueGlassBlock", email = "blueglassblock@outlook.com" },
@@ -8,10 +8,9 @@ authors = [
     { name = "Redlnn", email = "w731347477@gmail.com" },
 ]
 dependencies = [
-    "launart>=0.8.1",
-    "loguru>=0.6.0",
-    "playwright>=1.36.0",
-    "typing-extensions>=4.7.1",
+    "launart>=0.8.2",
+    "loguru>=0.7.2",
+    "playwright>=1.40.0",
 ]
 keywords = ["graia", "graiax", "launart", "playwright"]
 requires-python = ">=3.10"
@@ -38,11 +37,10 @@ build-backend = "pdm.backend"
 [tool.pdm]
 [tool.pdm.dev-dependencies]
 dev = [
-    "babel>=2.12.1",
-    "black>=23.7.0",
+    "babel>=2.14.0",
+    "black>=23.12.0",
     "creart>=0.3.0",
-    "isort>=5.12.0",
-    "richuru>=0.1.1",
+    "isort>=5.13.2",
 ]
 
 [tool.pdm.build]
