+++ src/graiax/playwright/__init__.py
@@ -1,3 +1,1 @@
-from .interface import PlaywrightBrowser as PlaywrightBrowser
-from .interface import PlaywrightContext as PlaywrightContext
 from .service import PlaywrightService as PlaywrightService
