+++ src/graiax/playwright/locale/zh_CN/LC_MESSAGES/graiax-playwright.po
@@ -6,9 +6,9 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: graiax-playwright 1.0.0\n"
+"Project-Id-Version: graiax-playwright 0.4.0\n"
 "Report-Msgid-Bugs-To: redlnn@graiax.cn\n"
-"POT-Creation-Date: 2023-03-26 17:35+0800\n"
+"POT-Creation-Date: 2023-12-17 16:58+0800\n"
 "PO-Revision-Date: 2023-03-19 01:06+0800\n"
 "Last-Translator: Red_lnn <redlnn@graiax.cn>\n"
 "Language: zh_Hans_CN\n"
@@ -17,41 +17,41 @@ msgstr ""
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel 2.12.1\n"
+"Generated-By: Babel 2.14.0\n"
 
-#: src/graiax/playwright/installer.py:29
+#: src/graiax/playwright/installer.py:30
 msgid ""
 "Start download Playwright for {browser_type} with dependencies, may "
 "require administrator privileges from you."
 msgstr "开始下载/检查更新所选的 Playwright 浏览器：{browser_type}，可能会向你请求[magenta]管理员权限[/]"
 
-#: src/graiax/playwright/installer.py:37
+#: src/graiax/playwright/installer.py:38
 msgid ""
 "Start download Playwright for {browser_type} with dependencies, may "
 "require you to access sudo."
 msgstr "开始下载/检查更新所选的 Playwright 浏览器：{browser_type}，可能会向你请求访问 [magenta]sudo[/]"
 
-#: src/graiax/playwright/installer.py:43
+#: src/graiax/playwright/installer.py:44
 msgid "Start download Playwright for {browser_type}."
 msgstr "开始下载/检查更新所选的 Playwright 浏览器：{browser_type}。"
 
-#: src/graiax/playwright/installer.py:62
+#: src/graiax/playwright/installer.py:63
 msgid "Downloaded [cyan]{file}[/] to [magenta]{path}[/]"
 msgstr "已下载 [cyan]{file}[/] 到 [magenta]{path}[/]"
 
-#: src/graiax/playwright/installer.py:66
+#: src/graiax/playwright/installer.py:67
 msgid "Download Failed:\n"
 msgstr "下载失败：\n"
 
-#: src/graiax/playwright/installer.py:70
+#: src/graiax/playwright/installer.py:71
 msgid "Failed to download Playwright for {browser_type}."
 msgstr "无法下载所选的 Playwright 浏览器：{browser_type}。"
 
-#: src/graiax/playwright/installer.py:71
+#: src/graiax/playwright/installer.py:72
 msgid "Please see: [magenta]https://playwright.dev/python/docs/intro[/]"
 msgstr "点击右侧链接查看更多: [magenta]https://playwright.dev/python/docs/intro[/]"
 
-#: src/graiax/playwright/installer.py:74
+#: src/graiax/playwright/installer.py:75
 msgid ""
 "Run [magenta]poetry run playwright install[/] or [magenta]pdm run "
 "playwright install[/] to install Playwright manually."
@@ -59,21 +59,32 @@ msgstr ""
 "运行 [magenta]poetry run playwright install[/] 或 [magenta]pdm run "
 "playwright install[/] 来手动安装 Playwright。"
 
-#: src/graiax/playwright/installer.py:80
+#: src/graiax/playwright/installer.py:81
 msgid "Playwright for {browser_type} is installed."
 msgstr "基于 {browser_type} 的 Playwright 已安装。"
 
-#: src/graiax/playwright/service.py:159
+#: src/graiax/playwright/service.py:174 src/graiax/playwright/service.py:198
+#: src/graiax/playwright/service.py:347 src/graiax/playwright/service.py:359
 msgid ""
-"Playwright service is launched by using a persistent context. Fetching "
-"browser is not supported."
-msgstr "Playwright 服务是通过使用持久上下文（Persistent Context）的方式来启动的。不支持获取 Browser 实例。"
+"Playwright has not been started yet, you cannot use the this method at "
+"this time"
+msgstr "Playwright 此时还未启动，你现在不可以使用此方法。"
 
-#: src/graiax/playwright/service.py:194
+#: src/graiax/playwright/service.py:177 src/graiax/playwright/service.py:350
+msgid ""
+"Playwright service is launched by using a persistent context. So you must"
+" use global context."
+msgstr "Playwright 服务是通过使用持久上下文（Persistent Context）的方式来启动的。所以你必须使用全局上下文。"
+
+#: src/graiax/playwright/service.py:181 src/graiax/playwright/service.py:354
+msgid "`Prsistent Context` cannot accept additional parameters. Ignore it."
+msgstr "持久性上下文（Prsistent Context）不支持额外参数，已忽略。"
+
+#: src/graiax/playwright/service.py:572
 msgid "Playwright is currently starting in persistent context mode."
 msgstr "Playwright 当前正以持久上下文模式启动。"
 
-#: src/graiax/playwright/service.py:199
+#: src/graiax/playwright/service.py:580
 msgid ""
 "Unable to launch Playwright for {browser_type}, please check the log "
 "output for the reason of failure. It is possible that some system "
@@ -83,6 +94,7 @@ msgstr ""
 "无法启动基于 {browser_type} 的 Playwright，请检查日志输出以获取失败原因。有可能是缺少一些系统依赖，你可以将 "
 "[magenta]`install_with_deps`[/] 设置为 [magenta]`True`[/] 以在下载浏览器时一同安装依赖"
 
-#: src/graiax/playwright/service.py:209
+#: src/graiax/playwright/service.py:590
 msgid "Playwright for {browser_type} is started."
 msgstr "基于 {browser_type} 的 Playwright 已启动。"
+
